pyproject.toml
[project]
name = "open-skills-host"
version = "0.1.0"
description = "安全、隔离、即插即用的 Agent Skill 执行宿主"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "strictyaml>=1.7.3",
    "langchain>=0.3.25",
    "langchain-core>=1.2",
    "langchain-openai>=1.1",
    "langgraph>=1.1",
    "python-dotenv>=1.1.0",
    "pydantic>=2.0.0",
    "rich>=14.3.3",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "httpx>=0.27.0",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
]
